+++ desktop/src/renderer/src/layouts/AppLayout.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from "@/hooks";
 import { Header } from "@bcgov/design-system-react-components";
 import { Button, Grid2 as Grid } from "@mui/material";
 import { ToastContainer } from "react-toastify";
-import { SideNav } from "../SideNav";
+import { SideNav } from "../components/SideNav";
 
 export const AppLayout = ({ children }) => {
 	const { navigate } = useNavigate();
